+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Feature: Money objects can be multiplied and divided.
 * Feature: Money can be converted using Exchange and ExchangeRate.  
 * Feature: Currencies can be looked up by ISO code, country code. 
+* Feature: Money objects can stringified with flexible formats, instead of just the currency's default format. 
 * Refactor: Most interfaces now return Result<T, MoneyError> instead of panicking or returning <T> (breaking change)
 * Refactor: Money::new now accepts i64 minor units instead of Decimal (breaking change)
 * Refactor: Currency::find takes strs instead of strings (breaking change)